error - possible loss of precision solved!
git-svn-id: file:///tmp/svnroot_only_rec/ReC/trunk@61 8c4c7310-a39f-45c4-bd0d-b10b2e5d99a4

+++ src/pt/utl/ist/elab/driver/serial/stamp/gamma/translators/StampConfigTranslator.java
@@ -39,7 +39,7 @@ public class StampConfigTranslator extends AbstractStampTranslator
         
         int vol_to_send = 350- ((volume - 5 ) * 350)/15;
         
-        int T = (1 / ((Integer)command.getCommandData(FREQ_STR)).intValue()/5.4253E-04);
+        int T = (int)(1. / ((Integer)command.getCommandData(FREQ_STR)).doubleValue()/5.4253E-04);
         
         String commandStr = command.getCommandIdentifier() + " " + vol_to_send + " " +  T + " " + numsamples;
         command.setCommand(commandStr);
